虚拟主播（开源版本）/requirements.txt
aiohttp==3.10.5
aiohttp==3.12.13
aiohttp==3.9.3
anthropic==0.57.1
bilibili_live==0.10.4
blive==0.2.5
edge_tts==7.0.0
edge_tts==7.0.2
Flask==3.1.1
httpx==0.28.1
openai==1.95.1
protobuf==4.25.3
pygame==2.6.1
PyYAML==6.0.2
PyYAML==6.0.1
PyYAML==6.0.2
qrcode==8.2
Requests==2.32.4
six==1.16.0
six==1.17.0
urllib3==2.5.0
websockets==14.2
